+++ archsim/src/arch/risc-v/RiscVComplianceEmulationModel.cpp
@@ -70,7 +70,7 @@ bool RiscVComplianceEmulationModel::InvokeSignal(int signum, uint32_t next_pc, S
 	return false;
 }
 
-archsim::abi::ExceptionAction RiscVComplianceEmulationModel::HandleException(archsim::core::thread::ThreadInstance *cpu, unsigned int category, unsigned int data)
+archsim::abi::ExceptionAction RiscVComplianceEmulationModel::HandleException(archsim::core::thread::ThreadInstance *cpu, uint64_t category, uint64_t data)
 {
 	if(category == 1024) {
 		GetSystem().GetPubSub().Publish(PubSubType::L1ICacheFlush, (void*)(uint64_t)0);
